Initial commit
It doesn't work as intended. Namely we get into in
unexplained infinite loop.

The ideal is to modify program to be more polymorphic,
by replacing (Map a (ACTree out a ))) with an appropriate
MapLike class, and then check if the program will work
with

[ (a, ACTree out a) ]

instead of (Map a (ACTree out a)).