[ACProfunctorTree] Initial commit. Working version of KMPalg.
So far this is just modified copy of KMPTree with slightly changed
types and corrected bugs.

PS. In KMPTree is incorrect function match and incorrect function
tie. The problem is that the both function require to distinguish
to states: when we didn't start to match and when we started to
match.